refactor: download command
